hop_count is set to be uint64_t

+++ dynamics_headers/particle_1d/iesh_particle_1d.hpp
@@ -255,7 +255,7 @@ namespace {
 				double nuclear_fric;
 				int Ndtq;
                 const double thermal_tau, thermal_tau_inv;
-                int hop_count;
+                uint64_t hop_count;
 		};
 
 };
