+++ runIESH_mpi.cpp
@@ -65,7 +65,7 @@ int main(int argc, char** argv)
     vector<double> sumel_Ep(Nrecord, 0.0);
 
     // evolution
-    int hop_count(0);
+    uint64_t hop_count(0);
     for (int ijob(0); ijob < Njob; ++ijob) {
         int j(mybatch[ijob]);
         ptcl_t ptcl(x[j], v[j], para.mass, para.kT, para.nuclear_fric, Nele, Nhole, para.Ndtq, para.thermal_tau, hamiltonian);
